+++ tournament-organiser-web/package.json
@@ -9,7 +9,7 @@
     "preview": "vite preview",
     "test:unit": "vitest",
     "test:e2e": "start-server-and-test preview http://localhost:4173 'cypress run --e2e --browser chrome'",
-    "test:e2e:open": "start-server-and-test preview http://localhost:4173 'cypress open --browser chrome'",
+    "test:e2e:open": "CYPRESS_BASE_URL=http://localhost:5173 cypress open --browser chrome",
     "test:e2e:ci": "CYPRESS_BASE_URL=http://localhost:3000 cypress run --e2e --browser chrome",
     "test:e2e:ci:open": "CYPRESS_BASE_URL=http://localhost:3000 cypress open --e2e",
     "test:e2e:dev": "start-server-and-test 'vite dev --port 4173' http://localhost:4173 'cypress open --e2e'",
