+++ tournament-organiser-web/src/components/UserLoginModal.vue
