+++ tournament-organiser-web/src/components/ui/FormInput.vue
@@ -1,6 +1,7 @@
 <template>
   <div class="flex flex-col gap-1">
     <input
+      :name="name"
       :type="type"
       :placeholder="placeholder"
       :value="value"
